Add files via upload
The project is based on the optimization of patient flow which encompasses quick, efficient and effective meeting of the demand for care by moving patients through care pathways while improving coordination of care, patient safety and health outcomes. We have implemented this project by using the Ford-Fulkerson algorithm which is a graph algorithm used to find the maximum flow in a flow network. It operates by repeatedly finding augmenting paths and updating the flow along those paths. The algorithm terminates when no more augmenting paths can be found. To keep things simple, we represented the graph as a 2D matrix for the implementation of Ford-Fulkerson algorithm. The code is in Java Programming where we have used various java packages such as java.io.*, java.lang.*, java.util.* and java.util.LinkedList. We created a visited array and mark all vertices as not visited then created a queue by using Java Collections Framework. Then applied a standard Breadth First Search Loop to check for the connection to the sink node. To handle runtime errors, the java exception handling is used. After the application of the project, we are able to reorganize patient pathways,improve patient experience,improve quality of care and reduce wait time of the patients in the hospitals.